Don't keep macOS struct fields on iOS
